requirements.txt
annotated-types==0.7.0
anyio==4.5.2
awsebcli==3.21.0
beautifulsoup4==4.12.3
blessed==1.20.0
blinker==1.9.0
botocore==1.35.60
cement==2.10.14
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.0
jmespath==1.0.1
MarkupSafe==3.0.2
openai==1.54.1
packaging==24.2
pathspec==0.10.1
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
semantic-version==2.10.0
setuptools==75.5.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
termcolor==2.4.0
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==1.26.20
wcwidth==0.2.13
Werkzeug==3.1.3
zipp==3.20.2
